// server/types/userProfile.ts
export interface UserProfileInfo {
  id: number;
  nickname: string;
  avatarUrl: string;
  status: number;
  createdAt: string;
  updatedAt: string;
  pointsBalance: number;
  uploadCount: number;
  downloadCount: number;
  likeCount: number;
  downloadedByOthersCount: number;
  likedByOthersCount: number;
}